Change the port the node server listens to
When the application is booted, BrowserSync proxies the server, using
port 3000 to serve the proxied site. The node server also starts up on
poort 3000, so the node server throws an error:

```Error: listen EADDRINUSE :::3000```

Changing the node server port to use something else (8080 in this
case) avoids this conflict. The BrowserSync config is updated to match.